Allow bash completions installation via make
By default, this uses the output from `pkg-config`, so the bash
completions directory (where surfraw's completion script will be placed)
is likely to be in `/usr/share/bash-completion/completions/`.

Perhaps the default should be in
`$datadir/bash-completion/completions/`? It would certainly make a lot
more sense since bash seems to be load completions from `/usr/local/...`
even when it is installed under `/usr/`.