Uniformly handle options before and after elvis determined
This removes the discrepancy between the behaviour of options before and
after the requested elvis is determined (i.e., `$elvis` is non-empty).
The code is also less likely to go out of sync with itself now.

The `-version` option was also fixed by this, allowing it to be used
plainly as in `sr -version`.