Launch SURFRAW_browser and BROWSER using `exec` call
This fixes an issue when using cli browsers under tmux and screen. When
the script launches a browser, tmux/screen display the pane as "sh"
since the browser is only a child of the script.

More generally, with launcher scripts, it's typically a good idea to
call the launch command via an `exec`. This makes the wrapper
"trasparent" to /proc while also saving a process.