Use datadir and `sr` when installing bash completions
This allows completions to be loaded on demand.  This doesn't use
`pkg-config`, which is what the `bash-completion` package recommends in
its README.

FYI, completions in `/etc/bash_completion.d/` are "only present for
backwards compatibility" [1].

The install-data-hook and uninstall-local rules are getting rather
complex.  Perhaps some automake magic could be employed to address this
(e.g, dist_ prefixes).

[1]: README.md file of https://github.com/scop/bash-completion