Provide richer completions
Each elvis can now specify its completions.  This should be agnostic to
shells--it just needs to have a completion script written for it.

This sources surfraw, so it needs to handle being run in a login shell.
Otherwise, `w3_argv0` would execute `basename -bash` after expansion,
and `basename` would try to parse `-bash` as the options `-b`, `-a`,
etc.