Provide richer completions
Each elvis can now specify its completions.  This should be agnostic to
shells--it just needs to have a completion script written for it.

This is still a bit rough around the edges but I think it works awesome.
Removing the space after completing options taking arguments (e.g.,
-browser=) would be great, but it's rather hard with this method.

The "copout" completion doesn't seem to work like in the old completion
script :/.  I'll fix this later... hopefully.