Add custom completions for each elvis
I tended to follow the options documented in the help output.  Some were
so big that I didn't want to do them fully, others were outdated, and
some diverged from what the code actually did.

As for `scpan`, I wonder why the env var for `-search=` is
`SURFRAW_scpan_opts` instead of `SURFRAW_scpan_search`, which would be
the usual.  There's no semantic difference since nothing special is done
to that value.  Maybe for ease of remembering?

The `W` elvis uses the default bash completer since it shouldn't
complete random words.  It's essentially an alias for `$SURFRAW_browser
$some_url`.

As for `aur`, its completions for `-language=` were taken from the
languages mentioned explicitly by the code.  I don't know what any of
the other supported languages are, if any.

The values of `-distro=` for `debcontents` and `debpackages` are
outdated, but it's better to have completions than none at all.  See #15
for updating this.