scpan: Remove comment calling for more docs
I don't know why I thought it was unclearly documented: It's right there
in the help text!  The values are just title-cased for style.

I wonder why the env var for `SURFRAW_scpan_search` is
`SURFRAW_scpan_opts` instead of the former, which would be the usual.
There's no semantic difference since nothing special is done to that
value.  Maybe for ease of remembering?