acronym: Update URL for acronym attic
Looks like the main site and the attic site have the same search
engine interface, so they both use the same query parameters.

I didn't compress the code into a variable for the base site URL (main
site or attic) to which "~/search/af.aspx?Acronym=" is appended so
that any future URL changes are simpler.