ntrs: Update for new search interface
I used mkelvis again here, but mkelvis wasn't as expressive as it
should have been.

I'd *expect* sites to treat `PARAM=` the same as the lack of `PARAM`
in the query string, but this site evidently does not.

It would also be convenient to allow escaping special characters to
our option syntax.