test: bugzilla: fix tests for new versions
A version of bugzilla introduced a new page for the ticket listing which
doesn't include the "bugs found" string anymore that was used as a
search string before.
Instead, look for the string that includes the content being searched to
make sure the elvi is working.

All other bugzilla tests still work due to the old page being served.

Signed-off-by: Thomas Böhler <witcher@wiredspace.de>